+++ FE/UI/src/app/components/about_us_component/about-us/about-us.component.ts
@@ -4,6 +4,6 @@ import { Component } from "@angular/core";
 @Component({
   selector: "app-about-us",
   templateUrl: "./about-us.component.html",
-  styleUrls: ["./about-us.component.scss"] // UI\src\app\components\about_us_component\about-us\about-us.component.scss
+  styleUrls: ["./about-us.component.scss"]
 })
 export class AboutUsComponent {}
