+++ FE/UI/src/app/components/header_component/header/header.component.ts
@@ -1,7 +1,7 @@
 // UI\src\app\components\header_component\header\header.component.ts
 
 import { Component } from "@angular/core";
-import { CartService } from "../../../services/cart.service";
+import { CartService } from "../../cart_page_component/cart-page/cart.service";
 import { Router, NavigationEnd } from "@angular/router";
 import { filter } from "rxjs/operators";
 import { SearchService } from "../../../services/search.service"; // Импортиране на SearchService
